requirements.txt
mockito==1.1.0
matplotlib==2.0.2
numpy==1.14.3
pandas==0.22.0
path.py==10.3.1
pycodestyle==2.4.0
